requirements.txt
requests==2.31.0
python-dotenv==1.0.1
google-genai
pydantic
fastapi
uvicorn
google-cloud-vision>=3.7.0
python-multipart>=0.0.6
pillow>=10.0.0
pdf2image>=1.16.0
